requirements_llm_agent.txt

# Python Version: 3.8+

# Core Web Framework
flask>=2.2.0                    # Web framework for API and frontend
werkzeug>=2.2.0                 # WSGI utility library (Flask dependency)

# Data Manipulation and Analysis
pandas>=2.0.0                   # Data manipulation and analysis
numpy>=1.24.0                   # Numerical computing
pyarrow>=10.0.0                 # Fast data storage (Feather format support)

# LangChain and AI
langchain-core>=0.1.0           # Core LangChain functionality
langchain-google-genai>=0.0.6   # Google Gemini AI integration
google-generativeai>=0.3.0      # Google Generative AI SDK

# Visualization
plotly>=5.17.0                  # Interactive charts (bar, line, pie, scatter)
kaleido>=0.2.1                  # Static image export for Plotly (PNG generation)

# PDF Generation
fpdf2>=2.7.0                    # PDF report generation with charts


python-dotenv>=0.21.0           # Environment variable management (.env files)

# Machine Learning and NLP
torch>=2.0.0                    # PyTorch for deep learning
transformers>=4.33.0            # Hugging Face transformers (BERT models)
statsmodels>=0.14.0             # Statistical models (ARIMA forecasting)
scikit-learn>=1.3.0             # Machine learning utilities
scipy>=1.10.0                   # Scientific computing

# Date and Time Handling
python-dateutil>=2.8.0          # Advanced date parsing and manipulation

# Excel File Support
openpyxl>=3.0.0                 # Excel file reading/writing (.xlsx format)
xlsxwriter>=3.0.0               # Alternative Excel writer (optional but recommended)

# Database Support
pymongo>=4.0.0                  # MongoDB database driver
motor>=3.0.0                    # Async MongoDB driver (optional)

# Task Scheduling
apscheduler>=3.10.0             # Background task scheduling (data refresh)

# Additional Utilities
requests>=2.28.0                # HTTP library (for API calls)
urllib3>=1.26.0                 # HTTP client (requests dependency)

# Optional but Recommended
# pillow>=9.0.0                 # Image processing (if needed for charts)
# matplotlib>=3.5.0             # Alternative plotting library
# seaborn>=0.12.0               # Statistical data visualization

# Development Dependencies (Optional)
# pytest>=7.0.0                 # Testing framework
# black>=22.0.0                 # Code formatter
# flake8>=5.0.0                 # Code linter
# mypy>=0.990                   # Type checker

